Add CdoGrid class to detect CDO objects
